api/requirements.txt
fastapi
uvicorn
pydantic
torch
transformers
langchain
langchain-community
sentence-transformers
faiss-cpu
pymongo
accelerate
huggingface_hub
python-dotenv


